Completed update functionality, did tests

+++ app/Http/Controllers/UserController.php
@@ -23,14 +23,16 @@ class UserController extends Controller
         return view('users.index', ['users' => $model->paginate(15)]);
     }
 
-    public function create(){
+    public function create()
+    {
         return view('users.create');
     }
 
-    public function store(NewUserRequest $request){
+    public function store(NewUserRequest $request)
+    {
         $role_id = Role::where('name', $request->role)->first()->id ?? null;
         $role = getModelFromRoleName($request->role) ?? null;
-        if(!$role){
+        if (!$role) {
             return redirect()->back()->withErrors(['role' => 'Invalid role']);
         }
         if (Gate::allows('admin', $role)) {
@@ -48,23 +50,34 @@ class UserController extends Controller
         }
     }
 
-    public function edit(User $user){
+    public function edit(User $user)
+    {
         return view('users.create', ['user' => $user]);
     }
 
-    public function destroy(User $user){
+    public function destroy(User $user)
+    {
+        if (auth()->user()->cannot('edit-delete-user', $user)) {
+            return redirect(route('users.index'))->withErrors(['role' => 'You may not delete this user']);
+        }
         $user->delete();
         return redirect()->route('users.index')->withStatus(__('User successfully deleted.'));
     }
-    public function update(UpdateUserRequest $request, User $user){
+
+    public function update(UpdateUserRequest $request, User $user)
+    {
         $role_id = Role::where('name', $request->role)->first()->id ?? null;
+        if (auth()->user()->cannot('edit-delete-user', $user)) {
+            return redirect(route('users.index'))->withErrors(['role' => 'You may not edit this user']);
+        }
         $user->user()->delete();
         User::where('id', $user->id)
             ->update([
-            'name' => $request->name,
-            'user_type' => getModelFromRoleName($request->role),
-            'user_id' => getModelFromRoleName($request->role)::create()->id,
-            'role_id' => $role_id,
+            'name' => $request->name ?? $user->name,
+            'email' => $request->email ?? $user->email,
+            'user_type' => getModelFromRoleName($request->role) ?? $user->user_type,
+            'user_id' => getModelFromRoleName($request->role)::create()->id ?? $user->user_id,
+            'role_id' => $role_id ?? $user->role_id,
         ]);
         return redirect()->route('users.index')->withStatus(__('User successfully updated.'));
     }
